+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Download current and previous versions from [https://github.com/ludus-leonis/Fre
 * fixed default piece backsides counting as side in library
 * fixed piece.h default handling (now equals w, not 0)
 * fixed browser caching artwork from previous template in same room
+* fixed user preferences to have global default values and ttl
 
 ## v0.12.0 - Hexing Heron